requirements.txt
streamlit
camel-ai
transformers
huggingface_hub
pymupdf
qdrant-client